backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
python-multipart>=0.0.9
requests>=2.31.0
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.7.0
httpx>=0.24.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
Pillow>=9.0.0
pytest>=8.0.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
pymongo==4.5.0
passlib>=1.7.4
tzdata>=2024.2
motor==3.3.1
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
pandas>=2.2.0
numpy>=1.26.0
jq>=1.6.0
typer>=0.9.0